moved AbstractRecorder to org.palladiosimulator.recorderframework.core
