+++ bundles/org.palladiosimulator.recorderframework.core/src/main/java/org/palladiosimulator/recorderframework/core/config/IRecorderConfiguration.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.recorderframework.config;
+package org.palladiosimulator.recorderframework.core.config;
 
 import java.util.Map;
