+++ bundles/org.palladiosimulator.recorderframework.edp2/src/org/palladiosimulator/recorderframework/edp2/config/AbstractEDP2RecorderConfigurationFactory.java
@@ -17,7 +17,7 @@ import org.palladiosimulator.metricspec.MetricDescription;
 import org.palladiosimulator.metricspec.MetricSetDescription;
 import org.palladiosimulator.recorderframework.config.AbstractRecorderConfiguration;
 import org.palladiosimulator.recorderframework.config.AbstractRecorderConfigurationFactory;
-import org.palladiosimulator.recorderframework.config.IRecorderConfiguration;
+import org.palladiosimulator.recorderframework.core.config.IRecorderConfiguration;
 import org.palladiosimulator.recorderframework.edp2.MeasuringTypeEquivalenceHelper;
 
 import de.uka.ipd.sdq.identifier.Identifier;
