+++ bundles/org.palladiosimulator.recorderframework.sensorframework/src/org/palladiosimulator/recorderframework/sensorframework/strategies/AbstractWriteDataStrategy.java
@@ -1,6 +1,6 @@
 package org.palladiosimulator.recorderframework.sensorframework.strategies;
 
-import org.palladiosimulator.recorderframework.config.IRecorderConfiguration;
+import org.palladiosimulator.recorderframework.core.config.IRecorderConfiguration;
 import org.palladiosimulator.recorderframework.sensorframework.SensorFrameworkRecorderConfiguration;
 import org.palladiosimulator.recorderframework.sensorframework.SensorHelper;
