+++ bundles/org.palladiosimulator.recorderframework.sensorframework/src/org/palladiosimulator/recorderframework/sensorframework/strategies/IWriteDataStrategy.java
@@ -1,7 +1,7 @@
 package org.palladiosimulator.recorderframework.sensorframework.strategies;
 
 import org.palladiosimulator.measurementframework.measureprovider.IMeasureProvider;
-import org.palladiosimulator.recorderframework.config.IRecorderConfiguration;
+import org.palladiosimulator.recorderframework.core.config.IRecorderConfiguration;
 
 /**
  * @deprecated Superseded by EDP2.
