+++ bundles/org.palladiosimulator.recorderframework.sensorframework/src/org/palladiosimulator/recorderframework/sensorframework/strategies/StateOfPassiveResourceWriteDataStrategy.java
@@ -9,7 +9,7 @@ import javax.measure.unit.SI;
 
 import org.palladiosimulator.measurementframework.measureprovider.IMeasureProvider;
 import org.palladiosimulator.metricspec.constants.MetricDescriptionConstants;
-import org.palladiosimulator.recorderframework.config.IRecorderConfiguration;
+import org.palladiosimulator.recorderframework.core.config.IRecorderConfiguration;
 import org.palladiosimulator.recorderframework.sensorframework.SensorFrameworkRecorderConfiguration;
 import org.palladiosimulator.recorderframework.sensorframework.SensorHelper;
