+++ bundles/org.palladiosimulator.recorderframework/META-INF/MANIFEST.MF
@@ -13,7 +13,8 @@ Require-Bundle: org.jscience;bundle-version="4.1.3",
  org.palladiosimulator.measurementframework;bundle-version="1.0.0",
  org.palladiosimulator.metricspec;bundle-version="1.0.0",
  org.palladiosimulator.edp2,
- org.palladiosimulator.commons
+ org.palladiosimulator.commons,
+ org.palladiosimulator.recorderframework.core;bundle-version="5.2.1";visibility:=reexport
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Export-Package: org.palladiosimulator.recorderframework,
