+++ bundles/org.palladiosimulator.recorderframework/src/org/palladiosimulator/recorderframework/config/AbstractRecorderConfiguration.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import org.palladiosimulator.commons.datastructureutils.MapHelper;
 import org.palladiosimulator.edp2.models.measuringpoint.MeasuringPoint;
 import org.palladiosimulator.metricspec.MetricDescription;
+import org.palladiosimulator.recorderframework.core.config.IRecorderConfiguration;
 
 /**
  * Abstract implementation for recorder configuration objects.
