+++ bundles/org.palladiosimulator.recorderframework/src/org/palladiosimulator/recorderframework/config/IRecorderConfigurationFactory.java
@@ -2,6 +2,8 @@ package org.palladiosimulator.recorderframework.config;
 
 import java.util.Map;
 
+import org.palladiosimulator.recorderframework.core.config.IRecorderConfiguration;
+
 /**
  * This interface specifies factory methods for creating configuration objects of recorders.
  *
